backend/requirements.txt
flask
flask-bcrypt
flask-sqlalchemy
flask-jwt-extended
flask-migrate
schedule
flask-CORS
firebase-admin
python-dotenv
authlib